+++ sdks/java/io/amazon-web-services2/src/test/java/org/apache/beam/sdk/io/aws2/kinesis/TestHelpers.java
@@ -44,7 +44,7 @@ import software.amazon.awssdk.services.kinesis.model.ListShardsResponse;
 import software.amazon.awssdk.services.kinesis.model.Record;
 import software.amazon.awssdk.services.kinesis.model.Shard;
 
-class Helpers {
+class TestHelpers {
   static final int SHARD_EVENTS = 100;
 
   static void mockShards(KinesisClient client, int count) {
